modified:   radix-4-2D-clean.cl 	new file:   radix-4-2D-clean.cl~ 	new file:   radix-4-2D-modi.cl 	new file:   radix-4-2D-modi.cl~ 	modified:   radix-4-2D.cl 	new file:   radix-4-2D.cl~ 	modified:   radix-4-big-clean.cl 	new file:   radix-4-big-clean.cl~ 	modified:   radix-4-big.cl 	new file:   radix-4-big.cl~ 	modified:   radix-4-float.cl 	new file:   radix-4-float.cl~ 	modified:   radix-4-init.cl 	new file:   radix-4-init.cl~ 	modified:   radix-4-interm.cl 	new file:   radix-4-interm.cl~ 	modified:   test_cluster.c 	modified:   vec_add.c 	modified:   vec_add.cl

+++ vec_add.cl
@@ -1,9 +1,9 @@
 //#pragma OPENCL EXTENSION cl_amd_fp64 : enable
 
 __kernel void sum(
-    __global const float *a,
-    __global const float *b, 
-    __global float *c,
+    __global const float16 *a,
+    __global const float16 *b, 
+    __global float16 *c,
     float a_mult,
     float b_mult,
     int n)
